// src/models/postgresql/payment-orders-model/enums.ts
export enum PaymentProviderTypes {
  PAYME = "payme",
  UZUM = "uzum",
  CLICK = "click",
}

export enum PaymentOrderPurposeTypes {
  CARD_TOPUP = "card_topup",
}

export enum PaymentOrderStatusTypes {
  PENDING = "pending",
  PROCESSING = "processing",
  PAID = "paid",
  CANCELLED = "cancelled",
  EXPIRED = "expired",
}